+++ pkg/handler/component.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"fmt"
