+++ pkg/handler/request-handler.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"os"
